Paths are working now
New buttons (first, last)
all buttons works perfectly

+++ Serializer/Serializer/Person.cs
@@ -9,10 +9,10 @@ namespace Serializer
     class Person
     {
         [NonSerialized] int id;
-        string name;
-        string address;
-        int phone;
-        DateTime dateOfRecording;
+        public string name;
+        public string address;
+        public int phone;
+        public DateTime dateOfRecording;
 
 
         public Person(string name, string address, int phone)
@@ -24,43 +24,21 @@ namespace Serializer
             dateOfRecording = DateTime.Now;
         }
 
-        public int GetId()
-        {
-            return id;
-        }
-
         public void Serialize(string output)
         {
-            if (File.Exists(output))
-            {
-                File.Delete(output);
-            }
-            // Create file to save the data
-            FileStream filestream = new FileStream(output, FileMode.Create);
-
-            // Create and use a BinaryFormatter object to perform the serialization
+            if (File.Exists(output)) File.Delete(output);
             BinaryFormatter formatter = new BinaryFormatter();
-            try
-            {
-                formatter.Serialize(filestream, this);
-            }
-            catch (SerializationException e)
-            {
-                Console.WriteLine("Failed to serialize. Reason: " + e.Message);
-                throw;
-            }
-
-            // Close the file
-            filestream.Close();
+            FileStream filestream = new FileStream(output, FileMode.Create, FileAccess.Write, FileShare.None);
+            formatter.Serialize(filestream, this);
         }
 
-        public override string ToString()
+        public static Person Deserialize(string fileNameToCall)
         {
-            return "id: " + id + "\n" +
-                   "name: " + name + "\n" +
-                   "address: " + address + "\n" +
-                   "phone: " + phone + "\n" +
-                   "Date of recording: " + dateOfRecording;
+            Stream filestream = new FileStream(fileNameToCall, FileMode.Open);
+            BinaryFormatter formatter = new BinaryFormatter();
+            Person person = (Person)formatter.Deserialize(filestream);
+            filestream.Close();
+            return person;
         }
     }
 }
